# docs/swagger.yaml
basePath: /cs/v1
definitions:
  course-service_internal_domain.Assignment:
    properties:
      deadline_days:
        type: integer
      description:
        type: string
      item_id:
        type: integer
      max_score:
        type: integer
    type: object
...
  course-service_internal_domain.Bank:
    properties:
      course_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  course-service_internal_domain.BankAnswer:
    properties:
      answer_text:
        type: string
      id:
        type: integer
      is_correct:
        type: boolean
      question_id:
        type: integer
    type: object
  course-service_internal_domain.BankQuestion:
    properties:
      answers:
        items:
          $ref: '#/definitions/course-service_internal_domain.BankAnswer'
        type: array
      bank_id:
        type: integer
      created_at:
        type: string
      default_points:
        type: integer
      id:
        type: integer
      question_text:
        type: string
      question_type:
        $ref: '#/definitions/course-service_internal_domain.QuestionType'
    type: object
...
  course-service_internal_domain.QuestionType:
    enum:
    - 0
    - 1
    - 2
    format: int32
    type: integer
    x-enum-varnames:
    - QuestionTypeSingleChoice
    - QuestionTypeMultipleChoice
    - QuestionTypeText
...
  course-service_internal_transport_http_v1_request.BankAnswer:
    properties:
      answer_text:
        type: string
      is_correct:
        type: boolean
    type: object
  course-service_internal_transport_http_v1_request.BulkCreateBankQuestion:
    properties:
      questions:
        items:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.CreateBankQuestion'
        type: array
    type: object
  course-service_internal_transport_http_v1_request.BulkUpdateBankQuestion:
    properties:
      questions:
        items:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.UpdateBankQuestion'
        type: array
    type: object
  course-service_internal_transport_http_v1_request.CreateAssignment:
    properties:
      deadline_days:
        type: integer
      description:
        type: string
      item_id:
        type: integer
      max_score:
        type: integer
    type: object
  course-service_internal_transport_http_v1_request.CreateBank:
    properties:
      course_id:
        type: integer
      description:
        type: string
      title:
        type: string
    type: object
  course-service_internal_transport_http_v1_request.CreateBankQuestion:
    properties:
      answers:
        items:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.BankAnswer'
        type: array
      bank_id:
        type: integer
      default_points:
        type: integer
      question_text:
        type: string
      question_type:
        type: integer
    type: object
...
  course-service_internal_transport_http_v1_request.UpdateAssignment:
    properties:
      deadline_days:
        type: integer
      description:
        type: string
      item_id:
        type: integer
      max_score:
        type: integer
    type: object
  course-service_internal_transport_http_v1_request.UpdateBank:
    properties:
      course_id:
        type: integer
      description:
        type: string
      title:
        type: string
    type: object
  course-service_internal_transport_http_v1_request.UpdateBankQuestion:
    properties:
      answers:
        items:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.BankAnswer'
        type: array
      bank_id:
        type: integer
      default_points:
        type: integer
      question_text:
        type: string
      question_type:
        type: integer
    type: object
...
  internal_transport_http_v1.ErrorResponse:
    properties:
      message:
        example: bad request
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    REST API for courses, sections, assignments, question banks and tests.
    Tests are also available under `/api` (same handlers as `/cs/v1`).
  title: Course Service API
  version: "1.0"
paths:
  /assignments:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/course-service_internal_domain.Assignment'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: List assignments
      tags:
      - assignments
    post:
      consumes:
      - application/json
      parameters:
      - description: Assignment
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.CreateAssignment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/course-service_internal_domain.Assignment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Create assignment
      tags:
      - assignments
  /assignments/{item_id}:
    delete:
      parameters:
      - description: Section item ID
        in: path
        name: item_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Delete assignment
      tags:
      - assignments
    get:
      parameters:
      - description: Section item ID
        in: path
        name: item_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course-service_internal_domain.Assignment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Get assignment
      tags:
      - assignments
    put:
      consumes:
      - application/json
      parameters:
      - description: Section item ID
        in: path
        name: item_id
        required: true
        type: integer
      - description: Assignment
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.UpdateAssignment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course-service_internal_domain.Assignment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Update assignment
      tags:
      - assignments
...
  /bank_questions:
    get:
      parameters:
      - description: Bank ID
        in: query
        name: bank_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/course-service_internal_domain.BankQuestion'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: List bank questions
      tags:
      - bank-questions
  /bank_questions/{id}/bulk:
    put:
      consumes:
      - application/json
      parameters:
      - description: Bank ID
        in: path
        name: id
        required: true
        type: integer
      - description: Questions
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.BulkUpdateBankQuestion'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Replace all questions in a bank
      tags:
      - bank-questions
  /bank_questions/bulk:
    post:
      consumes:
      - application/json
      parameters:
      - description: Questions
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.BulkCreateBankQuestion'
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Bulk create bank questions
      tags:
      - bank-questions
...
  /banks:
    get:
      parameters:
      - description: Course ID
        in: query
        name: course_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/course-service_internal_domain.Bank'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: List question banks
      tags:
      - banks
    post:
      consumes:
      - application/json
      parameters:
      - description: Bank
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.CreateBank'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/course-service_internal_domain.Bank'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Create question bank
      tags:
      - banks
  /banks/{id}:
    delete:
      parameters:
      - description: Bank ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Delete question bank
      tags:
      - banks
    put:
      consumes:
      - application/json
      parameters:
      - description: Bank ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bank
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.UpdateBank'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course-service_internal_domain.Bank'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Update question bank
      tags:
      - banks
...
schemes:
- http
swagger: "2.0"